# SimBlog17API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SimBlog17API",
    "version": "1.0.0",
    "description": "Lightweight forum backend API.\nSupports JWT access/refresh tokens, articles, comments, interactions, homepage aggregation, and admin moderation.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notification"
        ],
        "summary": "List my notifications",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageNotificationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "tags": [
          "Notification"
        ],
        "summary": "Get unread notification count",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseLong"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "put": {
        "tags": [
          "Notification"
        ],
        "summary": "Mark notification as read",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseNotificationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/read-all": {
      "put": {
        "tags": [
          "Notification"
        ],
        "summary": "Mark all notifications as read",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseLong"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - missing/invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 5,
              "statusMessage": "Unauthorized",
              "timeStamp": 1778002000000,
              "traceId": "5f8a6ca2-b8bf-4ad4-b3d4-0d7f7dd6a20b"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 200
        }
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string",
          "example": "publishedTime,desc"
        }
      },
...
    },
    "schemas": {
...
      "NotificationResponse": {
        "type": "object",
        "example": {
          "id": 3001,
          "type": "MODERATION",
          "targetType": "ARTICLE",
          "targetId": 42,
          "title": "文章审核结果",
          "message": "你的文章已审核通过并发布",
          "status": 0,
          "createTime": "2026-05-05 11:20:00",
          "readTime": null
        },
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "targetType": {
            "type": "string"
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "readTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "PageNotificationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "ApiStatusBase": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "statusMessage": {
            "type": "string"
          },
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ApiStatusResponseLong": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseNotificationResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/NotificationResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponsePageNotificationResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageNotificationResponse"
              }
            }
          }
        ]
      }
    }
  }
}